chore: add render updater bootstrap
