+++ apps/desktop/layer/main/src/updater/render/index.ts
@@ -1,0 +1,2 @@
+export type { RenderManifest } from "./bootstrap"
+export { CanUpdateRenderState, RenderUpdateBootstrapper, renderUpdater } from "./bootstrap"
