+++ apps/desktop/layer/main/src/updater/render/logger.ts
@@ -1,0 +1,3 @@
+import log from "electron-log"
+
+export const logger = log.scope("hot-updater")
